Various bugs fixed

+++ Omnet++ Project/src/Transmitter.cc
@@ -24,7 +24,7 @@ void Transmitter::initialize()
     /*
      * buffer related variable initialization
      */
-    bufferSize = par("bufferSize") + 1;
+    bufferSize = (int)par("bufferSize") + 1;
     buffer = std::vector<PacketMsg*>(bufferSize);
     start_idx = 0;
     end_idx = 0;
